get rid of logger service, move code to index.js
